Develop

+++ Assets/MirrorMarketSDK/Runtime/Models/Request/Marketplace/CancelNftListOnMarketplaceRequest.cs
@@ -9,7 +9,7 @@ namespace MirrorworldSDK.Models
     {
         public string mint_address;
 
-        public float price;
+        public double price;
 
         public string auction_house;
     }
